Transferencia de Dados com AutoMapper e Dtos

+++ Back/src/Eventos.API/Controllers/EventosController.cs
@@ -3,12 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Eventos.Application.Interfaces;
-using Eventos.Domain;
 using Eventos.Persistence.Data;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-
+using Eventos.Application.Dtos;
 namespace Eventos.API.Controllers
 {
     [ApiController]
@@ -27,7 +26,7 @@ namespace Eventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if(eventos == null) return NotFound("Nenhum evento encontrado");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -45,7 +44,7 @@ namespace Eventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if(evento == null) return NotFound("Nenhum evento encontrado");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -63,7 +62,7 @@ namespace Eventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if(eventos == null) return NotFound("Nenhum evento encontrado");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -75,7 +74,7 @@ namespace Eventos.API.Controllers
             }
         }
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
@@ -92,7 +91,7 @@ namespace Eventos.API.Controllers
             }
         }
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
@@ -113,10 +112,13 @@ namespace Eventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if(evento == null) return NoContent();
+
                 if(await _eventoService.DeleteEvento(id))
                     return Ok("Deletado");
                 else
-                    return BadRequest("Evento não deletado");   
+                    throw new Exception("Ocorreu um problema não específicado ao tentar deletar o evento.");
             }
             catch (Exception ex)
             {
